Modal

+++ Gestion.Web/Models/AddItemViewModel.cs
@@ -13,8 +13,8 @@ namespace Gestion.Web.Models
         [Range(1, int.MaxValue, ErrorMessage = "You must select a product.")]
         public int ProductoId { get; set; }
 
-        [Range(0.0001, double.MaxValue, ErrorMessage = "The quantiy must be a positive number")]
-        public double Cantidad { get; set; }
+        [Range(0.0001,int.MaxValue, ErrorMessage = "The quantiy must be a positive number")]
+        public int Cantidad { get; set; }
 
         public IEnumerable<SelectListItem> Productos { get; set; }
